requirements.txt
streamlit==1.31.0
Pillow==10.2.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
matplotlib==3.8.2
scikit-learn==1.3.2
scipy
torch
torchvision 
gdown